+++ app/controllers/admin/sessions_controller.rb
@@ -2,7 +2,7 @@
 
 class Admin::SessionsController < Devise::SessionsController
   # before_action :configure_sign_in_params, only: [:create]
-
+ 
   # GET /resource/sign_in
   # def new
   #   super
