+++ app/controllers/application_controller.rb
@@ -1,21 +1,6 @@
 class ApplicationController < ActionController::Base
-  before_action :authenticate_customer!, except: [:top, :about]
-  before_action :configure_permitted_parameters, if: :devise_controller?
-  before_action :authenticate_admin!, if: :admin_url
-  def after_sign_in_path_for(resource)
-    root_path
-  end
-
-  def after_sign_out_path_for(resource)
-    new_customer_session_path
-  end
-  
+   before_action :authenticate_admin!, if: :admin_url
   def admin_url
     request.fullpath.include?("/admin")
-  end
-  protected
-
-  def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:sign_up, keys: [:first_name, :last_name, :first_name_kana, :last_name_kana, :postcode, :address, :is_deleted, :phone_number])
   end
 end
