+++ app/controllers/public/sessions_controller.rb
@@ -2,7 +2,20 @@
 
 class Public::SessionsController < Devise::SessionsController
   # before_action :configure_sign_in_params, only: [:create]
+  before_action :authenticate_customer!, except: [:top, :about]
+  before_action :configure_permitted_parameters, if: :devise_controller?
+  
+  def after_sign_in_path_for(resource)
+    root_path
+  end
 
+  def after_sign_out_path_for(resource)
+    new_customer_session_path
+  end
+  
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:first_name, :last_name, :first_name_kana, :last_name_kana, :postcode, :address, :is_deleted, :phone_number])
+  end
   # GET /resource/sign_in
   # def new
   #   super
