+++ app/javascript/packs/application.js
@@ -12,6 +12,8 @@ import "jquery";
 import "popper.js";
 import "bootstrap";
 import "../stylesheets/application" 
+import '@fortawesome/fontawesome-free/js/all'
+
 
 Rails.start()
 Turbolinks.start()
