+++ app/views/admin/homes/top.html.erb
@@ -1,3 +1,3 @@
-<%= render 'header' %>
+
 <h1>Admin::Homes#top</h1>
 <p>Find me in app/views/admin/homes/top.html.erb</p>
